Module cansat_stm32f446::app

The RTIC application module

Re-exports

pub use rtic::Monotonic as _;

Modules

blink

Software task

gps_irq

Hardware task

idle

Idle loop

init

Initialization function

log_nmea

Software task

monotonics

Holds static methods for each monotonic.

shared_resources 🔒

Structs

Local 🔒

Shared 🔒

__rtic_internal_Monotonics

Monotonics used by the system

__rtic_internal_blinkLocalResources

Local resources blink has access to

__rtic_internal_blink_Context

Execution context

__rtic_internal_blink_MicrosecMono_SpawnHandle

__rtic_internal_gps_irqSharedResources

Shared resources gps_irq has access to

__rtic_internal_gps_irq_Context

Execution context

__rtic_internal_idleLocalResources

Local resources idle has access to

__rtic_internal_idleSharedResources

Shared resources idle has access to

__rtic_internal_idle_Context

Execution context

__rtic_internal_init_Context

Execution context

__rtic_internal_log_nmeaSharedResources

Shared resources log_nmea has access to

__rtic_internal_log_nmea_Context

Execution context

__rtic_internal_log_nmea_MicrosecMono_SpawnHandle

Functions

EXTI0 🔒 ⚠

Interrupt handler to dispatch tasks at priority 1

TIM2 🔒 ⚠

USART3 🔒 ⚠

__rtic_internal_blink_MicrosecMono_spawn_after

Spawns the task after a set duration relative to the current time

__rtic_internal_blink_MicrosecMono_spawn_at

Spawns the task at a fixed time instant

__rtic_internal_blink_spawn

Spawns the task directly

__rtic_internal_log_nmea_MicrosecMono_spawn_after

Spawns the task after a set duration relative to the current time

__rtic_internal_log_nmea_MicrosecMono_spawn_at

Spawns the task at a fixed time instant

__rtic_internal_log_nmea_spawn

Spawns the task directly

blink 🔒

gps_irq 🔒

idle 🔒

init 🔒

User code end

log_nmea 🔒

Type Definitions

MicrosecMono 🔒

Rx3 🔒

User code from within the module

Tx3 🔒